extract server file (to expose app without strating)
